requirements.txt
ansiwrap==0.8.4
apache-libcloud==2.4.0
appdirs==1.4.3
appnope==0.1.0
asn1crypto==0.24.0
astroid==2.0.4
atomicwrites==1.3.0
attrs==18.2.0
backcall==0.1.0
black==18.9b0
bleach==3.0.2
cacheout
cachetools==3.0.0
catboost==0.11.1
certifi==2019.3.9
cffi==1.11.5
chardet==3.0.4
click==6.7
coreapi==2.3.3
coreschema==0.0.4
croniter==0.3.30
cryptography==2.4.2
cycler==0.10.0
decorator==4.3.0
defusedxml==0.5.0
diskcache
distro==1.4.0
Django==2.2.1
django-allauth==0.38.0
django-cors-headers==2.4.0
django-filter==2.0.0
djangorestframework==3.9.4
djangorestframework-jsonapi==2.6.0
docker==4.0.1
docutils==0.14
dodgy==0.1.9
drf-yasg==1.11.1
entrypoints==0.3
enum34==1.1.6
flex==6.13.2
future==0.17.1
google-api-core==1.5.2
google-auth==1.6.1
google-cloud==0.34.0
google-cloud-core==0.28.1
google-cloud-storage==1.13.0
google-resumable-media==0.3.1
googleapis-common-protos==1.5.5
GPUtil==1.4.0
gunicorn==19.9.0
holidays==0.9.8
idna==2.7
inflection==0.3.1
ipykernel==5.1.0
ipython==7.2.0
ipython-genutils==0.2.0
ipywidgets==7.4.2
isort==4.3.4
itypes==1.1.0
jedi==0.13.1
Jinja2==2.10.1
joblib==0.13.0
jsonfield==2.0.2
jsonpointer==1.14
jsonschema==2.6.0
jupyter-client==5.2.3
jupyter-core==4.4.0
jupyterlab==0.35.4
jupyterlab-server==0.2.0
kiwisolver==1.0.1
lazy-object-proxy==1.3.1
libgravatar==0.2.3
livereload==2.6.0
Markdown==3.0.1
MarkupSafe==1.1.0
matplotlib==3.0.3
mccabe==0.6.1
missingno==0.4.1
mistune==0.8.4
mkdocs==1.0.4
mkdocs-material==4.0.1
more-itertools==5.0.0
mysql-connector-python==8.0.13
nbconvert==5.5.0
nbformat==4.4.0
notebook==5.7.8
nteract-on-jupyter==2.0.7
numpy==1.15.4
oauthlib==3.0.0
openpyxl
ortools==6.9.5824
pandas==0.23.4
pandocfilters==1.4.2
papermill==1.0.0
parso==0.3.1
pep8-naming==0.4.1
pexpect==4.6.0
pickleshare==0.7.5
Pillow==6.0.0
pkginfo==1.5.0.1
pluggy==0.8.1
prometheus-client==0.5.0
prompt-toolkit==2.0.7
prospector==1.1.6.2
protobuf==3.6.1
psutil==5.5.1
ptyprocess==0.6.0
py==1.7.0
pyarrow==0.13.0
pyasn1==0.4.4
pyasn1-modules==0.2.2
pycodestyle==2.4.0
pycparser==2.19
pycryptodome==3.7.2
pydocstyle==3.0.0
pyflakes==1.6.0
Pygments==2.3.0
pylint==2.1.1
pylint-celery==0.3
pylint-django==2.0.2
pylint-flask==0.5
pylint-plugin-utils==0.4
pymdown-extensions==6.0
pyOpenSSL==18.0.0
pyparsing==2.3.1
pytest==4.2.0
pytest-django==3.4.8
python-dateutil==2.7.5
python3-openid==3.1.0
pytz==2018.9
PyYAML==5.1
pyzmq==17.1.2
readme-renderer==24.0
requests==2.22.0
requests-oauthlib==1.2.0
requests-toolbelt==0.9.1
requirements-detector==0.6
retry==0.9.2
rfc3987==1.3.8
rope==0.11.0
rsa==4.0
ruamel.yaml==0.15.81
scikit-learn==0.20.0
scipy==1.1.0
seaborn==0.9.0
Send2Trash==1.5.0
setoptconf==0.2.0
simplejson==3.16.0
six==1.12.0
snowballstemmer==1.2.1
sqlparse==0.3.0
strict-rfc3339==0.7
stripe
swagger-spec-validator==2.4.1
tables
tenacity==5.0.4
terminado==0.8.1
testpath==0.4.2
textwrap3==0.9.2
toml==0.10.0
tornado==5.1.1
tqdm==4.30.0
traitlets==4.3.2
twine==1.12.1
typed-ast==1.1.1
uritemplate==3.0.0
urllib3==1.25.3
validate-email==1.3
wcwidth==0.1.7
webencodings==0.5.1
websocket-client==0.56.0
widgetsnbextension==3.4.2
wrapt==1.10.11
xlrd